fix: postgres caps and bigint for view/like count
- Add SETUID/SETGID/CHOWN/DAC_OVERRIDE caps to soundwave-pg service
  (required for postgres to switch from root to postgres user with
  cap_drop: ALL and no-new-privileges:true)
- Change view_count/like_count from IntegerField to BigIntegerField
  (YouTube view counts can exceed 2^31 for popular videos)
- Add migration 0008_biginteger_view_like_count

+++ backend/audio/models.py
@@ -32,8 +32,8 @@ class Audio(models.Model):
     thumbnail_url = models.URLField(max_length=500, blank=True)
     published_date = models.DateTimeField()
     downloaded_date = models.DateTimeField(auto_now_add=True)
-    view_count = models.IntegerField(default=0)
-    like_count = models.IntegerField(default=0)
+    view_count = models.BigIntegerField(default=0)
+    like_count = models.BigIntegerField(default=0)
     audio_format = models.CharField(max_length=20, default='m4a')
     bitrate = models.IntegerField(null=True, blank=True, help_text="Bitrate in kbps")
